--- hub/src/main/java/com/nowayback/hub/application/exception/HubApplicationException.java
@@ -1,10 +1,0 @@
-package com.nowayback.hub.application.exception;
-
-import com.nowayback.common.exception.ErrorCode;
-import com.nowayback.common.exception.GlobalException;
-
-public class HubApplicationException extends GlobalException {
-    public HubApplicationException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
