--- hub/src/main/java/com/nowayback/hub/domain/exception/HubDomainErrorCode.java
@@ -1,32 +1,0 @@
-package com.nowayback.hub.domain.exception;
-
-import com.nowayback.common.exception.ErrorCode;
-import org.springframework.http.HttpStatus;
-
-public enum HubDomainErrorCode implements ErrorCode {
-    ;
-    private final String code;
-    private final String message;
-    private final HttpStatus httpStatus;
-
-    @Override
-    public String getCode() {
-        return code;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-
-    @Override
-    public HttpStatus getStatus() {
-        return httpStatus;
-    }
-
-    HubDomainErrorCode(String code, String message, HttpStatus httpStatus) {
-        this.code = code;
-        this.message = message;
-        this.httpStatus = httpStatus;
-    }
-}
