+++ hub/src/main/java/com/nowayback/hub/domain/hub/exception/HubDomainErrorCode.java
@@ -1,14 +1,10 @@
-package com.nowayback.hub.application.exception;
+package com.nowayback.hub.domain.hub.exception;
 
 import com.nowayback.common.exception.ErrorCode;
 import org.springframework.http.HttpStatus;
 
-public enum HubApplicationErrorCode implements ErrorCode {
-
-    HUB_NAME_ALREADY_EXISTS_EXCEPTION("HUB2001","이미 존재하는 허브 이름입니다.", HttpStatus.CONFLICT),
-    HUB_ADDRESS_ALREADY_EXISTS_EXCEPTION("HUB2002", "이미 존재하는 허브 주소입니다.", HttpStatus.CONFLICT),
-    HUB_NOT_FOUND_EXCEPTION("HUB2003", "허브를 찾을 수 없습니다.", HttpStatus.NOT_FOUND),;
-
+public enum HubDomainErrorCode implements ErrorCode {
+    ;
     private final String code;
     private final String message;
     private final HttpStatus httpStatus;
@@ -28,7 +24,7 @@ public enum HubApplicationErrorCode implements ErrorCode {
         return httpStatus;
     }
 
-    HubApplicationErrorCode(String code, String message, HttpStatus httpStatus) {
+    HubDomainErrorCode(String code, String message, HttpStatus httpStatus) {
         this.code = code;
         this.message = message;
         this.httpStatus = httpStatus;
