+++ hub/src/test/java/com/nowayback/hub/presentation/HubControllerTest.java
@@ -1,12 +1,13 @@
 package com.nowayback.hub.presentation;
 
-import com.nowayback.hub.application.HubService;
-import com.nowayback.hub.application.command.CreateHubCommand;
-import com.nowayback.hub.application.command.UpdateHubCommand;
-import com.nowayback.hub.application.dto.CreateHubResult;
-import com.nowayback.hub.application.dto.UpdateHubResult;
-import com.nowayback.hub.application.exception.HubApplicationException;
-import com.nowayback.hub.domain.entity.Hub;
+import com.nowayback.hub.application.hub.HubService;
+import com.nowayback.hub.application.hub.command.CreateHubCommand;
+import com.nowayback.hub.application.hub.command.UpdateHubCommand;
+import com.nowayback.hub.application.hub.dto.CreateHubResult;
+import com.nowayback.hub.application.hub.dto.UpdateHubResult;
+import com.nowayback.hub.application.hub.exception.HubApplicationException;
+import com.nowayback.hub.domain.hub.entity.Hub;
+import com.nowayback.hub.presentation.hub.HubController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -20,7 +21,7 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import static com.nowayback.hub.application.exception.HubApplicationErrorCode.*;
+import static com.nowayback.hub.application.hub.exception.HubApplicationErrorCode.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
